backend/requirements.txt
Django
djangorestframework
python-dotenv
langchain
langgraph
openai
psycopg2-binary